requirements.txt
Django>=5.2
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
django-filter>=24.1
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
